+++ src/momo_args.h
@@ -59,6 +59,10 @@ struct MomoArgs {
   int sora_spotlight_number = 0;
   int sora_port = -1;
   bool sora_simulcast = false;
+  boost::optional<bool> sora_data_channel_signaling;
+  int sora_data_channel_signaling_timeout = 180;
+  boost::optional<bool> sora_ignore_disconnect_websocket;
+  int sora_disconnect_wait_timeout = 5;
 
   std::string test_document_root;
   int test_port = 8080;