+++ src/rtc/momo_video_encoder_factory.cpp
@@ -33,8 +33,6 @@
 #include "hwenc_nvcodec/nvcodec_h264_encoder.h"
 #endif
 
-#include "h264_format.h"
-
 MomoVideoEncoderFactory::MomoVideoEncoderFactory(
     VideoCodecInfo::Type vp8_encoder,
     VideoCodecInfo::Type vp9_encoder,
@@ -88,14 +86,14 @@ MomoVideoEncoderFactory::GetSupportedFormats() const {
 
   // H264
   std::vector<webrtc::SdpVideoFormat> h264_codecs = {
-      CreateH264Format(webrtc::H264::kProfileBaseline, webrtc::H264::kLevel3_1,
-                       "1"),
-      CreateH264Format(webrtc::H264::kProfileBaseline, webrtc::H264::kLevel3_1,
-                       "0"),
-      CreateH264Format(webrtc::H264::kProfileConstrainedBaseline,
-                       webrtc::H264::kLevel3_1, "1"),
-      CreateH264Format(webrtc::H264::kProfileConstrainedBaseline,
-                       webrtc::H264::kLevel3_1, "0")};
+      CreateH264Format(webrtc::H264Profile::kProfileBaseline,
+                       webrtc::H264Level::kLevel3_1, "1"),
+      CreateH264Format(webrtc::H264Profile::kProfileBaseline,
+                       webrtc::H264Level::kLevel3_1, "0"),
+      CreateH264Format(webrtc::H264Profile::kProfileConstrainedBaseline,
+                       webrtc::H264Level::kLevel3_1, "1"),
+      CreateH264Format(webrtc::H264Profile::kProfileConstrainedBaseline,
+                       webrtc::H264Level::kLevel3_1, "0")};
 
   if (h264_encoder_ == VideoCodecInfo::Type::VideoToolbox) {
     // VideoToolbox の場合は video_encoder_factory_ から H264 を拾ってくる