+++ src/sora/sora_client.cpp
@@ -11,6 +11,7 @@
 #include "ssl_verifier.h"
 #include "url_parts.h"
 #include "util.h"
+#include "zlib_helper.h"
 
 bool SoraClient::ParseURL(URLParts& parts) const {
   std::string url = config_.signaling_url;
@@ -19,7 +20,6 @@ bool SoraClient::ParseURL(URLParts& parts) const {
     throw std::exception();
   }
 
-  std::string default_port;
   if (parts.scheme == "wss") {
     return true;
   } else if (parts.scheme == "ws") {
@@ -67,11 +67,58 @@ SoraClient::SoraClient(boost::asio::io_context& ioc,
 
 SoraClient::~SoraClient() {
   destructed_ = true;
+  // 一応閉じる努力はする
+  if (using_datachannel_ && dc_) {
+    webrtc::DataBuffer disconnect =
+        ConvertToDataBuffer("signaling", R"({"type":"disconnect"})");
+    dc_->Close(
+        disconnect, [dc = dc_]() {}, config_.disconnect_wait_timeout);
+    dc_ = nullptr;
+    std::this_thread::sleep_for(std::chrono::milliseconds(300));
+  }
   // ここで OnIceConnectionStateChange が呼ばれる
   connection_ = nullptr;
 }
 
+void SoraClient::Close(std::function<void()> on_close) {
+  auto dc = std::move(dc_);
+  dc_ = nullptr;
+  auto ws = std::move(ws_);
+  ws_ = nullptr;
+  auto connection = std::move(connection_);
+  connection_ = nullptr;
+
+  if (using_datachannel_ && ws) {
+    webrtc::DataBuffer disconnect =
+        ConvertToDataBuffer("signaling", R"({"type":"disconnect"})");
+    dc->Close(
+        disconnect,
+        [self = shared_from_this(), dc, connection, ws = std::move(ws),
+         on_close]() {
+          ws->Close(
+              [self, ws, on_close](boost::system::error_code) { on_close(); });
+        },
+        config_.disconnect_wait_timeout);
+  } else if (using_datachannel_ && !ws) {
+    webrtc::DataBuffer disconnect =
+        ConvertToDataBuffer("signaling", R"({"type":"disconnect"})");
+    dc->Close(
+        disconnect, [dc, connection, on_close]() { on_close(); },
+        config_.disconnect_wait_timeout);
+  } else if (!using_datachannel_ && ws) {
+    boost::json::value disconnect = {{"type", "disconnect"}};
+    ws->WriteText(boost::json::serialize(disconnect),
+                  [self = shared_from_this(), ws](boost::system::error_code,
+                                                  std::size_t) {});
+    ws->Close([self = shared_from_this(), ws,
+               on_close](boost::system::error_code) { on_close(); });
+  } else {
+    on_close();
+  }
+}
+
 void SoraClient::Reset() {
+  watchdog_.Disable();
   connection_ = nullptr;
 
   URLParts parts;
@@ -80,6 +127,8 @@ void SoraClient::Reset() {
   } else {
     ws_.reset(new Websocket(ioc_));
   }
+
+  dc_.reset(new SoraDataChannelOnAsio(ioc_, this));
 }
 
 void SoraClient::Connect() {
@@ -101,9 +150,12 @@ void SoraClient::ReconnectAfter() {
 }
 
 void SoraClient::OnWatchdogExpired() {
-  RTC_LOG(LS_INFO) << __FUNCTION__ << " reconnecting...:";
-  Reset();
-  Connect();
+  RTC_LOG(LS_INFO) << __FUNCTION__ << " closing...";
+  Close([this]() {
+    RTC_LOG(LS_INFO) << __FUNCTION__ << " closed and reconnecting...";
+    Reset();
+    Connect();
+  });
 }
 
 void SoraClient::OnConnect(boost::system::error_code ec) {
@@ -111,14 +163,17 @@ void SoraClient::OnConnect(boost::system::error_code ec) {
     ReconnectAfter();
     return MOMO_BOOST_ERROR(ec, "Handshake");
   }
+  RTC_LOG(LS_INFO) << __FUNCTION__ << " connected";
 
   DoRead();
   DoSendConnect();
 }
 void SoraClient::DoRead() {
-  ws_->Read(std::bind(&SoraClient::OnRead, shared_from_this(),
-                      std::placeholders::_1, std::placeholders::_2,
-                      std::placeholders::_3));
+  ws_->Read([self = shared_from_this(), ws = ws_](boost::system::error_code ec,
+                                                  std::size_t bytes_transferred,
+                                                  std::string text) {
+    self->OnRead(ec, bytes_transferred, std::move(text));
+  });
 }
 
 void SoraClient::DoSendConnect() {
@@ -185,6 +240,14 @@ void SoraClient::DoSendConnect() {
     }
   }
 
+  if (config_.data_channel_signaling) {
+    json_message["data_channel_signaling"] = *config_.data_channel_signaling;
+  }
+  if (config_.ignore_disconnect_websocket) {
+    json_message["ignore_disconnect_websocket"] =
+        *config_.ignore_disconnect_websocket;
+  }
+
   ws_->WriteText(boost::json::serialize(json_message));
 }
 void SoraClient::DoSendPong() {
@@ -194,12 +257,27 @@ void SoraClient::DoSendPong() {
 void SoraClient::DoSendPong(
     const rtc::scoped_refptr<const webrtc::RTCStatsReport>& report) {
   std::string stats = report->ToJson();
-  boost::json::value json_message = {{"type", "pong"}, {"stats", stats}};
-  std::string str = R"({"type":"pong","stats":)" + stats + "}";
-  ws_->WriteText(std::move(str));
+  if (dc_ && using_datachannel_ && dc_->IsOpen("stats")) {
+    // DataChannel が使える場合は type: stats で DataChannel に送る
+    std::string str = R"({"type":"stats","reports":)" + stats + "}";
+    SendDataChannel("stats", str);
+  } else if (ws_) {
+    std::string str = R"({"type":"pong","stats":)" + stats + "}";
+    ws_->WriteText(std::move(str));
+  }
+}
+void SoraClient::DoSendUpdate(const std::string& sdp, std::string type) {
+  boost::json::value json_message = {{"type", type}, {"sdp", sdp}};
+  if (dc_ && using_datachannel_ && dc_->IsOpen("signaling")) {
+    // DataChannel が使える場合は DataChannel に送る
+    SendDataChannel("signaling", boost::json::serialize(json_message));
+  } else if (ws_) {
+    ws_->WriteText(boost::json::serialize(json_message));
+  }
 }
 
-void SoraClient::CreatePeerFromConfig(boost::json::value jconfig) {
+std::shared_ptr<RTCConnection> SoraClient::CreateRTCConnection(
+    boost::json::value jconfig) {
   webrtc::PeerConnectionInterface::RTCConfiguration rtc_config;
   webrtc::PeerConnectionInterface::IceServers ice_servers;
 
@@ -227,17 +305,8 @@ void SoraClient::CreatePeerFromConfig(boost::json::value jconfig) {
   }
 #endif
 
-  connection_ = manager_->CreateConnection(rtc_config, this);
-}
-
-void SoraClient::Close() {
-  ws_->Close(std::bind(&SoraClient::OnClose, shared_from_this(),
-                       std::placeholders::_1));
-}
-
-void SoraClient::OnClose(boost::system::error_code ec) {
-  if (ec)
-    return MOMO_BOOST_ERROR(ec, "close");
+  manager_->AddDataManager(dc_);
+  return manager_->CreateConnection(rtc_config, this);
 }
 
 void SoraClient::OnRead(boost::system::error_code ec,
@@ -251,88 +320,138 @@ void SoraClient::OnRead(boost::system::error_code ec,
   if (ec == boost::asio::error::operation_aborted)
     return;
 
-  if (ec)
+  if (ec) {
+    // とりあえず WS や DC を閉じておいて、後で再接続が起きるようにする
+    ReconnectAfter();
+    Close([]() {});
     return MOMO_BOOST_ERROR(ec, "Read");
+  }
 
   RTC_LOG(LS_INFO) << __FUNCTION__ << ": text=" << text;
 
   auto json_message = boost::json::parse(text);
   const std::string type = json_message.at("type").as_string().c_str();
   if (type == "offer") {
-    CreatePeerFromConfig(json_message.at("config"));
-    const std::string sdp = json_message.at("sdp").as_string().c_str();
-
-    connection_->SetOffer(sdp, [this, json_message]() {
-      // simulcast では offer の setRemoteDescription が終わった後に
-      // トラックを追加する必要があるため、ここで初期化する
-      manager_->InitTracks(connection_.get());
-
-      if (config_.simulcast &&
-          json_message.as_object().count("encodings") != 0) {
-        std::vector<webrtc::RtpEncodingParameters> encoding_parameters;
-
-        // "encodings" キーの各内容を webrtc::RtpEncodingParameters に変換する
-        auto encodings_json = json_message.at("encodings").as_array();
-        for (auto v : encodings_json) {
-          auto p = v.as_object();
-          webrtc::RtpEncodingParameters params;
-          // absl::optional<uint32_t> ssrc;
-          // double bitrate_priority = kDefaultBitratePriority;
-          // enum class Priority { kVeryLow, kLow, kMedium, kHigh };
-          // Priority network_priority = Priority::kLow;
-          // absl::optional<int> max_bitrate_bps;
-          // absl::optional<int> min_bitrate_bps;
-          // absl::optional<double> max_framerate;
-          // absl::optional<int> num_temporal_layers;
-          // absl::optional<double> scale_resolution_down_by;
-          // bool active = true;
-          // std::string rid;
-          // bool adaptive_ptime = false;
-          params.rid = p["rid"].as_string().c_str();
-          if (p.count("maxBitrate") != 0) {
-            params.max_bitrate_bps = p["maxBitrate"].to_number<int>();
+    // Data Channel の圧縮されたデータが送られてくるラベルを覚えておく
+    {
+      auto it = json_message.as_object().find("data_channels");
+      if (it != json_message.as_object().end()) {
+        const auto& ar = it->value().as_array();
+        for (const auto& v : ar) {
+          if (v.at("compress").as_bool()) {
+            compressed_labels_.insert(v.at("label").as_string().c_str());
           }
-          if (p.count("minBitrate") != 0) {
-            params.min_bitrate_bps = p["minBitrate"].to_number<int>();
-          }
-          if (p.count("scaleResolutionDownBy") != 0) {
-            params.scale_resolution_down_by =
-                p["scaleResolutionDownBy"].to_number<double>();
-          }
-          if (p.count("maxFramerate") != 0) {
-            params.max_framerate = p["maxFramerate"].to_number<double>();
-          }
-          if (p.count("active") != 0) {
-            params.active = p["active"].as_bool();
-          }
-          if (p.count("adaptivePtime") != 0) {
-            params.adaptive_ptime = p["adaptivePtime"].as_bool();
-          }
-          encoding_parameters.push_back(params);
         }
-        connection_->SetEncodingParameters(std::move(encoding_parameters));
       }
+    }
 
-      connection_->CreateAnswer(
-          [this](webrtc::SessionDescriptionInterface* desc) {
-            std::string sdp;
-            desc->ToString(&sdp);
-            boost::json::value json_message = {{"type", "answer"},
-                                               {"sdp", sdp}};
-            ws_->WriteText(boost::json::serialize(json_message));
-          });
-    });
-  } else if (type == "update") {
+    connection_ = CreateRTCConnection(json_message.at("config"));
     const std::string sdp = json_message.at("sdp").as_string().c_str();
-    connection_->SetOffer(sdp, [this]() {
-      connection_->CreateAnswer(
-          [this](webrtc::SessionDescriptionInterface* desc) {
-            std::string sdp;
-            desc->ToString(&sdp);
-            boost::json::value json_message = {{"type", "update"},
-                                               {"sdp", sdp}};
-            ws_->WriteText(boost::json::serialize(json_message));
-          });
+
+    connection_->SetOffer(sdp, [self = shared_from_this(), json_message]() {
+      boost::asio::post(self->ioc_, [self, json_message]() {
+        if (!self->connection_) {
+          return;
+        }
+
+        // simulcast では offer の setRemoteDescription が終わった後に
+        // トラックを追加する必要があるため、ここで初期化する
+        self->manager_->InitTracks(self->connection_.get());
+
+        if (self->config_.simulcast &&
+            json_message.as_object().count("encodings") != 0) {
+          std::vector<webrtc::RtpEncodingParameters> encoding_parameters;
+
+          // "encodings" キーの各内容を webrtc::RtpEncodingParameters に変換する
+          auto encodings_json = json_message.at("encodings").as_array();
+          for (auto v : encodings_json) {
+            auto p = v.as_object();
+            webrtc::RtpEncodingParameters params;
+            // absl::optional<uint32_t> ssrc;
+            // double bitrate_priority = kDefaultBitratePriority;
+            // enum class Priority { kVeryLow, kLow, kMedium, kHigh };
+            // Priority network_priority = Priority::kLow;
+            // absl::optional<int> max_bitrate_bps;
+            // absl::optional<int> min_bitrate_bps;
+            // absl::optional<double> max_framerate;
+            // absl::optional<int> num_temporal_layers;
+            // absl::optional<double> scale_resolution_down_by;
+            // bool active = true;
+            // std::string rid;
+            // bool adaptive_ptime = false;
+            params.rid = p["rid"].as_string().c_str();
+            if (p.count("maxBitrate") != 0) {
+              params.max_bitrate_bps = p["maxBitrate"].to_number<int>();
+            }
+            if (p.count("minBitrate") != 0) {
+              params.min_bitrate_bps = p["minBitrate"].to_number<int>();
+            }
+            if (p.count("scaleResolutionDownBy") != 0) {
+              params.scale_resolution_down_by =
+                  p["scaleResolutionDownBy"].to_number<double>();
+            }
+            if (p.count("maxFramerate") != 0) {
+              params.max_framerate = p["maxFramerate"].to_number<double>();
+            }
+            if (p.count("active") != 0) {
+              params.active = p["active"].as_bool();
+            }
+            if (p.count("adaptivePtime") != 0) {
+              params.adaptive_ptime = p["adaptivePtime"].as_bool();
+            }
+            encoding_parameters.push_back(params);
+          }
+          self->connection_->SetEncodingParameters(
+              std::move(encoding_parameters));
+        }
+
+        self->connection_->CreateAnswer(
+            [self](webrtc::SessionDescriptionInterface* desc) {
+              std::string sdp;
+              desc->ToString(&sdp);
+
+              boost::asio::post(self->ioc_, [self, sdp]() {
+                if (!self->connection_) {
+                  return;
+                }
+
+                boost::json::value json_message = {{"type", "answer"},
+                                                   {"sdp", sdp}};
+                self->ws_->WriteText(boost::json::serialize(json_message));
+              });
+            });
+      });
+    });
+  } else if (type == "update" || type == "re-offer") {
+    if (connection_ == nullptr) {
+      return;
+    }
+    std::string answer_type = type == "update" ? "update" : "re-answer";
+    const std::string sdp = json_message.at("sdp").as_string().c_str();
+    connection_->SetOffer(sdp, [self = shared_from_this(), answer_type]() {
+      boost::asio::post(self->ioc_, [self, answer_type]() {
+        if (!self->connection_) {
+          return;
+        }
+
+        // エンコーディングパラメータの情報がクリアされるので設定し直す
+        if (self->config_.simulcast) {
+          self->connection_->ResetEncodingParameters();
+        }
+
+        self->connection_->CreateAnswer(
+            [self, answer_type](webrtc::SessionDescriptionInterface* desc) {
+              std::string sdp;
+              desc->ToString(&sdp);
+              boost::asio::post(self->ioc_, [self, sdp, answer_type]() {
+                if (!self->connection_) {
+                  return;
+                }
+
+                self->DoSendUpdate(sdp, answer_type);
+              });
+            });
+      });
     });
   } else if (type == "notify") {
     const std::string event_type =
@@ -363,15 +482,121 @@ void SoraClient::OnRead(boost::system::error_code ec,
     auto it = json_message.as_object().find("stats");
     if (it != json_message.as_object().end() && it->value().as_bool()) {
       connection_->GetStats(
-          [this](
+          [self = shared_from_this()](
               const rtc::scoped_refptr<const webrtc::RTCStatsReport>& report) {
-            DoSendPong(report);
+            self->DoSendPong(report);
           });
     } else {
       DoSendPong();
     }
+  } else if (type == "switched") {
+    // Data Channel による通信の開始
+    using_datachannel_ = true;
+
+    // ignore_disconnect_websocket == true の場合は WS を切断する
+    auto it = json_message.as_object().find("ignore_disconnect_websocket");
+    if (it != json_message.as_object().end() && it->value().as_bool() && ws_) {
+      RTC_LOG(LS_INFO) << "Close WebSocket for DataChannel";
+      auto ws = ws_;
+      ws_ = nullptr;
+      ws->Close([self = shared_from_this(), ws](boost::system::error_code) {});
+
+      watchdog_.Enable(config_.data_channel_signaling_timeout);
+      return;
+    }
   }
   DoRead();
+}
+
+webrtc::DataBuffer SoraClient::ConvertToDataBuffer(const std::string& label,
+                                                   const std::string& input) {
+  bool compressed = compressed_labels_.find(label) != compressed_labels_.end();
+  RTC_LOG(LS_INFO) << "Convert to DataBuffer: label=" << label
+                   << " compressed=" << compressed << " input=" << input;
+  const std::string& str = compressed ? ZlibHelper::Compress(input) : input;
+  return webrtc::DataBuffer(rtc::CopyOnWriteBuffer(str), compressed);
+}
+
+void SoraClient::SendDataChannel(const std::string& label,
+                                 const std::string& input) {
+  webrtc::DataBuffer data = ConvertToDataBuffer(label, input);
+  dc_->Send(label, data);
+}
+
+void SoraClient::OnStateChange(
+    rtc::scoped_refptr<webrtc::DataChannelInterface> data_channel) {}
+
+void SoraClient::OnMessage(
+    rtc::scoped_refptr<webrtc::DataChannelInterface> data_channel,
+    const webrtc::DataBuffer& buffer) {
+  if (!using_datachannel_ || !dc_) {
+    return;
+  }
+
+  std::string label = data_channel->label();
+  bool compressed = compressed_labels_.find(label) != compressed_labels_.end();
+  std::string data;
+  if (compressed) {
+    data = ZlibHelper::Uncompress(buffer.data.cdata(), buffer.size());
+  } else {
+    data.assign((const char*)buffer.data.cdata(),
+                (const char*)buffer.data.cdata() + buffer.size());
+  }
+
+  RTC_LOG(LS_INFO) << "label=" << label << " data=" << data;
+
+  // ハンドリングする必要のあるラベル以外は何もしない
+  if (label != "signaling" && label != "stats") {
+    return;
+  }
+
+  boost::json::error_code ec;
+  auto json = boost::json::parse(data, ec);
+  if (ec) {
+    RTC_LOG(LS_ERROR) << "JSON Parse Error ec=" << ec.message();
+    return;
+  }
+
+  watchdog_.Reset();
+
+  if (label == "signaling") {
+    const std::string type = json.at("type").as_string().c_str();
+    if (type == "re-offer") {
+      const std::string sdp = json.at("sdp").as_string().c_str();
+      connection_->SetOffer(sdp, [self = shared_from_this()]() {
+        boost::asio::post(self->ioc_, [self]() {
+          if (!self->connection_) {
+            return;
+          }
+
+          // エンコーディングパラメータの情報がクリアされるので設定し直す
+          if (self->config_.simulcast) {
+            self->connection_->ResetEncodingParameters();
+          }
+
+          self->connection_->CreateAnswer(
+              [self](webrtc::SessionDescriptionInterface* desc) {
+                std::string sdp;
+                desc->ToString(&sdp);
+                boost::asio::post(self->ioc_, [self, sdp]() {
+                  if (!self->connection_) {
+                    return;
+                  }
+                  self->DoSendUpdate(sdp, "re-answer");
+                });
+              });
+        });
+      });
+    }
+  }
+
+  if (label == "stats") {
+    connection_->GetStats(
+        [self = shared_from_this()](
+            const rtc::scoped_refptr<const webrtc::RTCStatsReport>& report) {
+          self->DoSendPong(report);
+        });
+  }
 }
 
 // WebRTC からのコールバック