+++ src/sora/sora_client.h
@@ -5,6 +5,7 @@
 #include <cstdlib>
 #include <functional>
 #include <memory>
+#include <set>
 #include <string>
 
 // Boost
@@ -16,6 +17,7 @@
 #include "metrics/stats_collector.h"
 #include "rtc/rtc_manager.h"
 #include "rtc/rtc_message_sender.h"
+#include "sora_data_channel_on_asio.h"
 #include "url_parts.h"
 #include "watchdog.h"
 #include "websocket.h"
@@ -38,11 +40,16 @@ struct SoraClientConfig {
   int spotlight_number = 0;
   int port = -1;
   bool simulcast = false;
+  boost::optional<bool> data_channel_signaling;
+  int data_channel_signaling_timeout = 180;
+  boost::optional<bool> ignore_disconnect_websocket;
+  int disconnect_wait_timeout = 5;
 };
 
 class SoraClient : public std::enable_shared_from_this<SoraClient>,
                    public RTCMessageSender,
-                   public StatsCollector {
+                   public StatsCollector,
+                   public SoraDataChannelObserver {
   SoraClient(boost::asio::io_context& ioc,
              RTCManager* manager,
              SoraClientConfig config);
@@ -55,18 +62,18 @@ class SoraClient : public std::enable_shared_from_this<SoraClient>,
         new SoraClient(ioc, manager, std::move(config)));
   }
   ~SoraClient();
+  void Close(std::function<void()> on_close);
 
   void Reset();
   void Connect();
-  void Close();
 
   webrtc::PeerConnectionInterface::IceConnectionState GetRTCConnectionState()
       const;
   std::shared_ptr<RTCConnection> GetRTCConnection() const;
 
-  void GetStats(
-      std::function<void(
-          const rtc::scoped_refptr<const webrtc::RTCStatsReport>&)> callback) override;
+  void GetStats(std::function<
+                void(const rtc::scoped_refptr<const webrtc::RTCStatsReport>&)>
+                    callback) override;
 
  private:
   void ReconnectAfter();
@@ -79,14 +86,27 @@ class SoraClient : public std::enable_shared_from_this<SoraClient>,
   void DoSendPong();
   void DoSendPong(
       const rtc::scoped_refptr<const webrtc::RTCStatsReport>& report);
-  void CreatePeerFromConfig(boost::json::value jconfig);
+  void DoSendUpdate(const std::string& sdp, std::string type);
+  std::shared_ptr<RTCConnection> CreateRTCConnection(
+      boost::json::value jconfig);
 
  private:
   void OnConnect(boost::system::error_code ec);
-  void OnClose(boost::system::error_code ec);
   void OnRead(boost::system::error_code ec,
               std::size_t bytes_transferred,
               std::string text);
+
+ private:
+  webrtc::DataBuffer ConvertToDataBuffer(const std::string& label,
+                                         const std::string& input);
+  void SendDataChannel(const std::string& label, const std::string& input);
+
+ private:
+  // DataChannel 周りのコールバック
+  void OnStateChange(
+      rtc::scoped_refptr<webrtc::DataChannelInterface> data_channel) override;
+  void OnMessage(rtc::scoped_refptr<webrtc::DataChannelInterface> data_channel,
+                 const webrtc::DataBuffer& buffer) override;
 
  private:
   // WebRTC からのコールバック
@@ -103,7 +123,10 @@ class SoraClient : public std::enable_shared_from_this<SoraClient>,
 
  private:
   boost::asio::io_context& ioc_;
-  std::unique_ptr<Websocket> ws_;
+  std::shared_ptr<Websocket> ws_;
+  std::shared_ptr<SoraDataChannelOnAsio> dc_;
+  bool using_datachannel_ = false;
+  std::set<std::string> compressed_labels_;
 
   std::atomic_bool destructed_ = {false};